Refactor: segregate core utilities and APIs into tdk module
- Moved all core functions and utilities from simulator to tdk module.
- Simulator now contains only the UI and mock service.
- Enables use of tdk as a standalone helper library without the simulator.

fix: Apply light mode theme to Register Vehicle Config UI
