+++ simulator/mockservices/chassis.py
@@ -24,15 +24,15 @@ import re
 from uprotocol.proto.umessage_pb2 import UMessage
 from uprotocol.transport.ulistener import UListener
 
-from simulator.core.abstract_service import BaseService
-from simulator.core.exceptions import ValidationError
-from simulator.target.protofiles.vehicle.chassis.v1.chassis_service_pb2 import (
+from simulator.utils.exceptions import ValidationError
+from tdk.core.abstract_service import BaseService
+from tdk.target.protofiles.vehicle.chassis.v1.chassis_service_pb2 import (
     UpdateTireRequest,
 )
-from simulator.target.protofiles.vehicle.chassis.v1.chassis_topics_pb2 import (
+from tdk.target.protofiles.vehicle.chassis.v1.chassis_topics_pb2 import (
     Tire,
 )
-from simulator.utils.constant import KEY_URI_PREFIX
+from tdk.utils.constant import KEY_URI_PREFIX
 
 
 class ChassisService(BaseService):
