+++ simulator/ui/static/js/sidebar.js
@@ -39,13 +39,14 @@ function resetViews(value) {
         document.getElementById('vehicle_config_div').hidden = true
         document.getElementById("configservicesdialog").hidden = true
         document.getElementById("someip_save").hidden = true
-        document.getElementById("register_ve_config").className = "card Services box-content-data config-firstRow"
+        document.getElementById("register_ve_config").classList.remove("config-someip");
+        document.getElementById("register_ve_config").classList.add("config-firstRow");
         if (value == "SOME/IP") {
             document.getElementById('someip_config_div').hidden = false
             document.getElementById("configservicesdialog").hidden = false
             document.getElementById("someip_save").hidden = false
-            document.getElementById("register_ve_config").className = "card Services box-content-data config-someip"
-
+            document.getElementById("register_ve_config").classList.remove("config-firstRow");
+            document.getElementById("register_ve_config").classList.add("config-someip");
 
         } else if (value == "ZENOH") {
             document.getElementById('zenoh_config_div').hidden = false
@@ -68,11 +69,27 @@ function resetViews(value) {
     } catch { }
 
     try {
-        console.log("neelam" + isFromRefresh)
-
-        //        if (!isFromRefresh) {
         closebtnClicked()
-        //        }
     } catch {
     }
 }
+
+// Function to initialize the select element with the saved value
+function initializeSelect() {
+    // Retrieve the previously saved value
+    const savedValue = localStorage.getItem('utransportConfig');
+
+    // Get the select element
+    const selectElement = document.getElementById('utransportConfig');
+
+    // Set the saved value or default to the first option if not found
+    if (savedValue && selectElement.querySelector(`option[value="${savedValue}"]`)) {
+        selectElement.value = savedValue;
+        setTransport(savedValue)
+    } else {
+        // Optionally handle the case where savedValue is not found or is invalid
+        selectElement.value = 'BINDER';  // Default value if no valid saved value
+    }
+}
+// Initialize the select element when the page loads
+window.onload = initializeSelect;
