+++ simulator/ui/templates/includes/sidebar.html
@@ -162,7 +162,7 @@
 
             <div class="sidebarFooter copyright-content">
                 <div>
-                    <span>Simulator Version - 0.1.0-dev</span> &nbsp
+                    <span>Simulator Version - 0.1.1-dev</span> &nbsp
                 </div>
                 <div>
                     <u>
