+++ tdk/core/protobuf_autoloader.py
@@ -1,21 +1,14 @@
 """
-SPDX-FileCopyrightText: Copyright (c) 2024 Contributors to the Eclipse Foundation
+SPDX-FileCopyrightText: 2024 Contributors to the Eclipse Foundation
 
 See the NOTICE file(s) distributed with this work for additional
 information regarding copyright ownership.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+This program and the accompanying materials are made available under the
+terms of the Apache License Version 2.0 which is available at
 
     http://www.apache.org/licenses/LICENSE-2.0
 
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-SPDX-FileType: SOURCE
 SPDX-License-Identifier: Apache-2.0
 """
 
@@ -32,9 +25,9 @@ from collections import defaultdict
 
 from google.protobuf.descriptor import FieldDescriptor
 
-from simulator.target import protofiles as proto
-from simulator.utils import constant
-from simulator.utils.constant import (
+from tdk.target import protofiles as proto
+from tdk.utils import constant
+from tdk.utils.constant import (
     RESOURCE_CATALOG_CSV_NAME,
     RESOURCE_CATALOG_JSON_NAME,
 )
