+++ tdk/helper/binder_utransport.py
@@ -1,21 +1,14 @@
 """
-SPDX-FileCopyrightText: Copyright (c) 2024 Contributors to the Eclipse Foundation
+SPDX-FileCopyrightText: 2024 Contributors to the Eclipse Foundation
 
 See the NOTICE file(s) distributed with this work for additional
 information regarding copyright ownership.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+This program and the accompanying materials are made available under the
+terms of the Apache License Version 2.0 which is available at
 
     http://www.apache.org/licenses/LICENSE-2.0
 
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-SPDX-FileType: SOURCE
 SPDX-License-Identifier: Apache-2.0
 """
 
@@ -46,7 +39,7 @@ from uprotocol.uri.validator.urivalidator import UriValidator
 from uprotocol.uuid.factory.uuidfactory import Factories
 from uprotocol.uuid.serializer.longuuidserializer import LongUuidSerializer
 
-from simulator.utils.constant import REPO_URL
+from tdk.utils.constant import REPO_URL
 
 # Dictionary to store requests
 m_requests = {}
@@ -274,7 +267,7 @@ class SocketClient:
         return self._rpc_request_callbacks
 
 
-class AndroidBinder(UTransport, RpcClient):
+class SocketTransport(UTransport, RpcClient):
     def __init__(self):
         self.client = SocketClient()
         # Start a separate thread for receiving
