+++ tdk/helper/transport_configuration.py
@@ -1,33 +1,20 @@
 """
-SPDX-FileCopyrightText: Copyright (c) 2024 Contributors to the Eclipse Foundation
+SPDX-FileCopyrightText: 2024 Contributors to the Eclipse Foundation
 
 See the NOTICE file(s) distributed with this work for additional
 information regarding copyright ownership.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+This program and the accompanying materials are made available under the
+terms of the Apache License Version 2.0 which is available at
 
     http://www.apache.org/licenses/LICENSE-2.0
 
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-SPDX-FileType: SOURCE
 SPDX-License-Identifier: Apache-2.0
 """
 
 import json
-from concurrent.futures import Future
 
-from uprotocol.proto.uattributes_pb2 import CallOptions
-from uprotocol.proto.umessage_pb2 import UMessage
-from uprotocol.proto.upayload_pb2 import UPayload
 from uprotocol.proto.uri_pb2 import UAuthority, UEntity, UUri
-from uprotocol.proto.ustatus_pb2 import UStatus
-from uprotocol.transport.ulistener import UListener
 
 try:
     from uprotocol_vsomeip.vsomeip_utransport import (
@@ -35,16 +22,16 @@ try:
         VsomeipTransport,
     )
 
-    from simulator.core.someip_helper import SomeipHelper
+    from tdk.transport.someip_helper import SomeipHelper
 except ImportError:
     pass
 
 from uprotocol.uri.factory.uresource_builder import UResourceBuilder
 
-from simulator.core.binder_utransport import AndroidBinder
+from tdk.helper.binder_utransport import SocketTransport
 
 
-class TransportLayer:
+class TransportConfiguration:
     _instance = None
     _initialized = False
 
@@ -57,45 +44,45 @@ class TransportLayer:
         if not self._initialized:
             # Initialize the singleton instance
             self._initialized = True
-            self.__instance = None
+            self.ut_instance = None
             self.__ZENOH_IP = "10.0.3.3"
             self.__ZENOH_PORT = 9090
             self.__SOMEIP_UNICAST_IP = "127.0.0.1"
             self.__SOMEIP_MULTICAST_IP = "224.224.224.245"
-            self.__utransport = "BINDER"
+            self.utransport = "BINDER"
             self._update_instance()
 
     def set_transport(self, transport: str):
-        if self.__utransport != transport:
+        if self.utransport != transport:
             print(
                 "set transport, previous is",
-                self.__utransport,
+                self.utransport,
                 "current is",
                 transport,
             )
-            self.__utransport = transport
+            self.utransport = transport
             self._update_instance()
 
     def get_transport(self):
-        return self.__utransport
+        return self.utransport
 
     def set_zenoh_config(self, ip, port):
-        self.__utransport = "ZENOH"
+        self.utransport = "ZENOH"
         if self.__ZENOH_PORT != port or self.__ZENOH_IP != ip:
             self.__ZENOH_PORT = port
             self.__ZENOH_IP = ip
             self._update_instance()
 
     def set_someip_config(self, localip, multicast):
-        self.__utransport = "SOME/IP"
+        self.utransport = "SOME/IP"
         self.__SOMEIP_UNICAST_IP = localip
         self.__SOMEIP_MULTICAST_IP = multicast
         self._update_instance()
 
     def _update_instance(self):
-        if self.__utransport == "BINDER":
-            self.__instance = AndroidBinder()
-        elif self.__utransport == "ZENOH":
+        if self.utransport == "BINDER":
+            self.ut_instance = SocketTransport()
+        elif self.utransport == "ZENOH":
             import zenoh
 
             zenoh_ip = self.__ZENOH_IP
@@ -108,13 +95,13 @@ class TransportLayer:
                 conf.insert_json5(zenoh.config.CONNECT_KEY, json.dumps(endpoint))
             from up_client_zenoh.upclientzenoh import UPClientZenoh
 
-            self.__instance = UPClientZenoh(
+            self.ut_instance = UPClientZenoh(
                 conf, UAuthority(name="simulator_authority"), UEntity(name="simulator_entity", version_major=1)
             )
 
-        elif self.__utransport == "SOME/IP":
+        elif self.utransport == "SOME/IP":
             self.__helper = VsomeipHelper()
-            self.__instance = VsomeipTransport(
+            self.ut_instance = VsomeipTransport(
                 helper=SomeipHelper(),
                 source=UUri(
                     authority=UAuthority(name="simulator_authority"),
@@ -124,27 +111,3 @@ class TransportLayer:
                 unicast=self.__SOMEIP_UNICAST_IP,
                 multicast=(self.__SOMEIP_MULTICAST_IP, 30490),
             )
-
-    def invoke_method(self, topic: UUri, payload: UPayload, calloptions: CallOptions) -> Future:
-        return self.__instance.invoke_method(topic, payload, calloptions)
-
-    def send(self, umessage: UMessage) -> UStatus:
-        return self.__instance.send(umessage)
-
-    def register_listener(self, topic: UUri, listener: UListener) -> UStatus:
-        return self.__instance.register_listener(topic, listener)
-
-    def unregister_listener(self, topic: UUri, listener: UListener) -> UStatus:
-        return self.__instance.unregister_listener(topic, listener)
-
-    def start_service(self, entity) -> bool:
-        if self.__utransport == "BINDER":
-            return self.__instance.start_service(entity)
-        else:
-            return True
-
-    def create_topic(self, entity, topics, listener):
-        if self.__utransport == "BINDER":
-            return self.__instance.create_topic(entity, topics, listener)
-        else:
-            return True
