+++ tdk/tools/generate_resource_catalog.py
@@ -1,21 +1,14 @@
 """
-SPDX-FileCopyrightText: Copyright (c) 2024 Contributors to the Eclipse Foundation
+SPDX-FileCopyrightText: 2024 Contributors to the Eclipse Foundation
 
 See the NOTICE file(s) distributed with this work for additional
 information regarding copyright ownership.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+This program and the accompanying materials are made available under the
+terms of the Apache License Version 2.0 which is available at
 
     http://www.apache.org/licenses/LICENSE-2.0
 
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-SPDX-FileType: SOURCE
 SPDX-License-Identifier: Apache-2.0
 """
 
@@ -27,8 +20,8 @@ import pkgutil
 
 from google._upb._message import RepeatedCompositeContainer
 
-from simulator.target import protofiles as proto
-from simulator.utils.constant import (
+from tdk.target import protofiles as proto
+from tdk.utils.constant import (
     KEY_URI_PREFIX,
     RESOURCE_CATALOG_CSV_NAME,
     RESOURCE_CATALOG_DIR,
