+++ tdk/tools/pull_and_compile_protos.py
@@ -1,21 +1,14 @@
 """
-SPDX-FileCopyrightText: Copyright (c) 2024 Contributors to the Eclipse Foundation
+SPDX-FileCopyrightText: 2024 Contributors to the Eclipse Foundation
 
 See the NOTICE file(s) distributed with this work for additional
 information regarding copyright ownership.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+This program and the accompanying materials are made available under the
+terms of the Apache License Version 2.0 which is available at
 
     http://www.apache.org/licenses/LICENSE-2.0
 
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-SPDX-FileType: SOURCE
 SPDX-License-Identifier: Apache-2.0
 """
 
@@ -27,7 +20,7 @@ import subprocess
 import git
 from git import Repo
 
-from simulator.utils.constant import (
+from tdk.utils.constant import (
     PROTO_OUTPUT_DIR,
     PROTO_REPO_DIR,
     REPO_URL,
@@ -101,27 +94,27 @@ def process_python_protofiles(directory):
                 replace_in_file(
                     file_path,
                     r"from vehicle",
-                    "from simulator.target.protofiles.vehicle",
+                    "from tdk.target.protofiles.vehicle",
                 )
                 replace_in_file(
                     file_path,
                     r"from example",
-                    "from simulator.target.protofiles.example",
+                    "from tdk.target.protofiles.example",
                 )
                 replace_in_file(
                     file_path,
                     r"from common",
-                    "from simulator.target.protofiles.common",
+                    "from tdk.target.protofiles.common",
                 )
                 replace_in_file(
                     file_path,
                     r"import uservices_options_pb2",
-                    "import simulator.target.protofiles.uservices_options_pb2",
+                    "import tdk.target.protofiles.uservices_options_pb2",
                 )
                 replace_in_file(
                     file_path,
                     r"import units_pb2",
-                    "import simulator.target.protofiles.units_pb2",
+                    "import tdk.target.protofiles.units_pb2",
                 )
                 replace_in_file(
                     file_path,
